Вынесены переменные в отдельный файл, добавлены модуля для записи в них классов
