+++ src/plan.txt
@@ -1,1 +1,13 @@
 Наш план
+Логика работы класса Popup и его дочерних классов:
+1)Создаём класс
+2)Класс попап принимает один параметр - селектор попапа, который нужно открыть, либо закрыть;
+3)Внутри мы его сохраняем в приватное поле. 
+4)Далее создаётся два публичных метода openPopup и closePopup, внутри соответственно открытие и закрытие попапов, также добавляем приватный метод _handleEscClose, который
+отвечает за закрытие по клику на esc.
+5)В конце создаём публичный метод для слушателей, которые отвечают за открытие и закрытие модальных окон.
+
+Создание дочерних классов:
+1)Создаём класс PopupWithImage (Открытие мордального окна карточки с подписью);
+2)В конструкторы принимается 2 параметра: (ссылка на картинку, подпись)
+3)Перезаписываем метод родителя, метод open должен при открытии модального окна подставлять значения из экзмпляра класса Card;
